fixing a base fail case in the page parser

+++ tools/lib.sh
@@ -206,7 +206,7 @@ pl_fallback() {
 get_page() {
   if [[ -s "$1/$PAGE" ]]; then
     curl -Ls $(resolve "$1") > "$tmp/$PAGE"
-    if [[ $( stat -c %s "$1/$PAGE" ) -lt $( stat -c %s "$tmp/$PAGE" ) ]]; then
+    if [[ -e "$tmp/$PAGE" && $( stat -c %s "$1/$PAGE" ) -lt $( stat -c %s "$tmp/$PAGE" ) ]]; then
       mv "$tmp/$PAGE" "$1/$PAGE"
     fi
   else
